pyproject.toml
[project]
name = "hexlet-price-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-vite>=3.1.0",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "inertia-django>=1.2.0",
    "ruff>=0.12.2",
]
